refactor: reservation으로 전부 변경
