+++ src/test/java/roomescape/repository/ReservationRepositoryTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import roomescape.domain.reservation.ReservationWithTimeAndTheme;
+import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.ReservationCommand;
 import roomescape.domain.reservationTime.ReservationTime;
 import roomescape.domain.theme.Theme;
@@ -21,7 +21,7 @@ import roomescape.repository.reservation.ReservationRepository;
 public class ReservationRepositoryTest extends BaseRepositoryTest {
     private ReservationRepository reservationRepository;
 
-    private final ReservationWithTimeAndTheme INIT_RESERVATION = new ReservationWithTimeAndTheme(
+    private final Reservation INIT_RESERVATION = new Reservation(
             1,
             "브라운",
             "2023-08-05",
@@ -50,7 +50,7 @@ public class ReservationRepositoryTest extends BaseRepositoryTest {
     @Test
     @DisplayName("존재하는 ID로 조회 시 해당 예약을 반환한다")
     void getReservationTest() {
-        Optional<ReservationWithTimeAndTheme> reservation = reservationRepository.getReservationWithTimeAndTheme(1);
+        Optional<Reservation> reservation = reservationRepository.getReservationWithTimeAndTheme(1);
 
         assertThat(reservation).isPresent();
         assertThat(reservation.get().name()).isEqualTo("브라운");
@@ -59,7 +59,7 @@ public class ReservationRepositoryTest extends BaseRepositoryTest {
     @Test
     @DisplayName("존재하지 않는 ID로 조회 시 empty를 반환한다")
     void getReservationEmptyTest() {
-        Optional<ReservationWithTimeAndTheme> reservation = reservationRepository.getReservationWithTimeAndTheme(999L);
+        Optional<Reservation> reservation = reservationRepository.getReservationWithTimeAndTheme(999L);
 
         assertThat(reservation).isEmpty();
     }
@@ -67,7 +67,7 @@ public class ReservationRepositoryTest extends BaseRepositoryTest {
     @Test
     @DisplayName("전체 예약 테스트 정상적으로 가져오는 지 테스트")
     void getAllReservationTest() {
-        List<ReservationWithTimeAndTheme> reservations = reservationRepository.getAllReservation(null);
+        List<Reservation> reservations = reservationRepository.getAllReservation(null);
 
         assertThat(reservations).containsExactly(INIT_RESERVATION);
     }
@@ -76,7 +76,7 @@ public class ReservationRepositoryTest extends BaseRepositoryTest {
     @DisplayName("예약 삭제 정상적으로 작동하는 지 테스트")
     void deleteReservationTest() {
         reservationRepository.deleteReservation(1);
-        List<ReservationWithTimeAndTheme> reservations = reservationRepository.getAllReservation(null);
+        List<Reservation> reservations = reservationRepository.getAllReservation(null);
 
         assertThat(reservations).isNotIn(INIT_RESERVATION);
     }
@@ -87,9 +87,9 @@ public class ReservationRepositoryTest extends BaseRepositoryTest {
         LocalDate date = LocalDate.now();
         reservationRepository.addReservation(new ReservationCommand("테스트", date, 1, 1));
 
-        List<ReservationWithTimeAndTheme> reservations = reservationRepository.getAllReservation(null);
+        List<Reservation> reservations = reservationRepository.getAllReservation(null);
 
-        ReservationWithTimeAndTheme expectedReservation = new ReservationWithTimeAndTheme(2, "테스트", date, new ReservationTime(1, LocalTime.parse("10:00")), new Theme(1, "테마1", "테마 설명", "image url"));
+        Reservation expectedReservation = new Reservation(2, "테스트", date, new ReservationTime(1, LocalTime.parse("10:00")), new Theme(1, "테마1", "테마 설명", "image url"));
 
         assertThat(reservations.size()).isEqualTo(2);
         assertThat(reservations).contains(expectedReservation);
@@ -104,11 +104,11 @@ public class ReservationRepositoryTest extends BaseRepositoryTest {
         insertReservation("테스트2", "2023-12-15", 1, 1);
 
 
-        List<ReservationWithTimeAndTheme> reservations = reservationRepository.getAllReservation("테스트");
+        List<Reservation> reservations = reservationRepository.getAllReservation("테스트");
 
-        List<ReservationWithTimeAndTheme> expectedReservation = List.of(
-                new ReservationWithTimeAndTheme(2, "테스트", "2023-09-15", new ReservationTime(1, LocalTime.parse("10:00")), new Theme(1, "테마1", "테마 설명", "image url")),
-                new ReservationWithTimeAndTheme(4, "테스트", "2023-11-15", new ReservationTime(1, LocalTime.parse("10:00")), new Theme(1, "테마1", "테마 설명", "image url"))
+        List<Reservation> expectedReservation = List.of(
+                new Reservation(2, "테스트", "2023-09-15", new ReservationTime(1, LocalTime.parse("10:00")), new Theme(1, "테마1", "테마 설명", "image url")),
+                new Reservation(4, "테스트", "2023-11-15", new ReservationTime(1, LocalTime.parse("10:00")), new Theme(1, "테마1", "테마 설명", "image url"))
         );
 
         assertThat(reservations.size()).isEqualTo(2);
@@ -120,8 +120,8 @@ public class ReservationRepositoryTest extends BaseRepositoryTest {
     void getAllReservationWithBlankNameTest() {
         insertReservation("테스트", "2023-09-15", 1, 1);
 
-        List<ReservationWithTimeAndTheme> resultWithEmpty = reservationRepository.getAllReservation("");
-        List<ReservationWithTimeAndTheme> resultWithBlank = reservationRepository.getAllReservation("   ");
+        List<Reservation> resultWithEmpty = reservationRepository.getAllReservation("");
+        List<Reservation> resultWithBlank = reservationRepository.getAllReservation("   ");
 
         assertThat(resultWithEmpty).hasSize(2);
         assertThat(resultWithBlank).hasSize(2);
@@ -171,12 +171,12 @@ public class ReservationRepositoryTest extends BaseRepositoryTest {
 
         assertThat(updatedCount).isEqualTo(1);
 
-        ReservationWithTimeAndTheme updatedInfo = reservationRepository.getReservationWithTimeAndTheme(1L).get();
+        Reservation updatedInfo = reservationRepository.getReservationWithTimeAndTheme(1L).get();
         assertAll(
                 () -> assertThat(updatedInfo.name()).isEqualTo("테스트"),
                 () -> assertThat(updatedInfo.date()).isEqualTo(date),
                 () -> assertThat(updatedInfo.time().id()).isEqualTo(2L),
-                () -> assertThat(updatedInfo.reservationTheme().id()).isEqualTo(2L)
+                () -> assertThat(updatedInfo.theme().id()).isEqualTo(2L)
         );
     }
 }
