+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -10,8 +10,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.dao.DataIntegrityViolationException;
 import roomescape.domain.reservation.ReservationCommand;
-import roomescape.domain.reservation.ReservationWithTime;
-import roomescape.domain.reservation.ReservationWithTimeAndTheme;
+import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservationTime.ReservationTime;
 import roomescape.domain.reservationTime.ReservationTimeCommand;
 import roomescape.domain.reservationTime.ReservationTimeCondition;
@@ -24,9 +23,8 @@ import roomescape.repository.reservationTime.ReservationTimeRepository;
 public class ReservationTimeServiceTest {
     private ReservationRepository createReservationRepository(boolean isUsedInReservation) {
         return new ReservationRepository() {
-            @Override public Optional<ReservationWithTimeAndTheme> getReservationWithTimeAndTheme(long id) { return Optional.empty(); }
-            @Override public Optional<ReservationWithTime> getReservationWithTime(long id) { return Optional.empty(); };
-            @Override public List<ReservationWithTimeAndTheme> getAllReservation(String name) { return List.of(); }
+            @Override public Optional<Reservation> getReservationWithTimeAndTheme(long id) { return Optional.empty(); }
+            @Override public List<Reservation> getAllReservation(String name) { return List.of(); }
             @Override public long addReservation(ReservationCommand reservationCommand) { return 0; }
             @Override public void deleteReservation(long id) {}
             @Override public int updateAll(long id, ReservationCommand cmd) { return 0; }
